third_party/WebKit/LayoutTests/fast/dom/HTMLScriptElement/append-child-adopt-node-crash-expected.txt

CONSOLE MESSAGE: line 16: PASS unless crash.
 
